feat: abstract route handling
