+++ docs/modules/client.rst
@@ -8,3 +8,5 @@ Client
 .. autoclass:: Client
     :members:
     :undoc-members:
+    :special-members: __init__
+    :inherited-members:
