+++ docs/modules/server.rst
@@ -7,3 +7,4 @@ Server
     :members:
     :undoc-members:
     :special-members: __init__
+    :inherited-members:
