+++ zonis/__init__.py
@@ -1,10 +1,12 @@
 from .packet import Packet, RequestPacket, ClientToServerPacket
 from .exceptions import *
+from .route_registration import RouteHandler
 from .router import Router
 from .server import Server
 from .client import Client, route
 
 __all__ = (
+    "RouteHandler",
     "Packet",
     "RequestPacket",
     "Server",
